Fix item damage bug. Fix aura show positon.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1354 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ItemPrototype.h
@@ -162,8 +162,8 @@ enum ITEM_SUBCLASS_PROJECTILE
 
 struct _Damage
 {
-    float DamageMin;
-    float DamageMax;
+    uint32 DamageMin;
+    uint32 DamageMax;
     uint32 DamageType;
 
 };
